move signup fetch to userActions and fix userSet
